Move files (the move implementation) one level up.
