+++ multiversxetl/config/worker_config_mainnet.json
